requirements.txt
# Core scientific computing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Machine Learning & Deep Learning
torch>=2.0.0
torchvision>=0.15.0
tensorflow>=2.13.0
scikit-learn>=1.3.0
scikit-image>=0.21.0

# Computer Vision
opencv-python>=4.8.0
Pillow>=10.0.0
albumentations>=1.3.0

# LLM Integration
openai>=1.0.0
langchain>=0.1.0
transformers>=4.35.0
accelerate>=0.24.0
sentence-transformers>=2.2.0

# Vector Database & Embeddings
chromadb>=0.4.0
faiss-cpu>=1.7.4

# Data Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# Experiment Tracking & MLOps
wandb>=0.16.0
mlflow>=2.9.0

# Configuration & Utilities
pyyaml>=6.0
python-dotenv>=1.0.0
tqdm>=4.66.0
loguru>=0.7.0

# Jupyter & Notebooks
jupyter>=1.0.0
notebook>=7.0.0
ipywidgets>=8.1.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Code Quality
black>=23.0.0
flake8>=6.1.0
isort>=5.12.0

# API & Web
fastapi>=0.104.0
uvicorn>=0.24.0
requests>=2.31.0

# Geospatial (for lunar imagery)
rasterio>=1.3.0
geopandas>=0.14.0

# Documentation
sphinx>=7.2.0
sphinx-rtd-theme>=2.0.0
